+++ src/components/modules/saver.ts
@@ -28,15 +28,11 @@ export default class Saver extends Module {
    * @returns {OutputData}
    */
   public async save(): Promise<OutputData> {
-    const { BlockManager, Tools, ModificationsObserver } = this.Editor;
+    const { BlockManager, Tools } = this.Editor;
     const blocks = BlockManager.blocks,
         chainData = [];
 
     try {
-      /**
-       * Disable onChange callback on save to not to spam those events
-       */
-      ModificationsObserver.disable();
 
       blocks.forEach((block: Block) => {
         chainData.push(this.getSavedData(block));
@@ -50,9 +46,7 @@ export default class Saver extends Module {
       return this.makeOutput(sanitizedData);
     } catch (e) {
       _.logLabeled(`Saving failed due to the Error %o`, 'error', e);
-    } finally {
-      ModificationsObserver.enable();
-    }
+    } 
   }
 
   /**